images/vertex_pipelines/requirements.txt
pytest==5.3.2
gunicorn==20.1.0
fastapi==0.73.0
uvicorn==0.17.0
uvloop==0.16.0
httptools==0.3.0
pydantic==1.9.0
kfp==1.8.11
google-cloud-aiplatform==1.10.0
google-cloud-pipeline-components==0.3.1
google-cloud-storage==1.44.0